+++ src/infra/database/sea_orm/models/payment.rs
@@ -26,7 +26,7 @@ pub struct Model {
     pub created_at: DateTime,
     pub updated_at: Option<DateTime>,
     pub btc_address: Option<String>,
-    pub btc_block_height: Option<i32>,
+    pub btc_block_height: Option<i64>,
     pub reserved_amount: i64,
     #[sea_orm(column_type = "JsonBinary", nullable)]
     pub raw_success_action: Option<Json>,
